small changes in avaliation till end and added random to base aval + better comparassion between Players

+++ includes/cpuplayer.h
@@ -28,7 +28,7 @@ public:
     ~BetterCpuPlayer() override = default;
     
     Coord choseSquare(const Game& game) override;
-    float avaliateMoveTillEnd(Coord move, std::unique_ptr<Game> game);
+    int avaliateMoveTillEnd(Coord move, std::unique_ptr<Game> game);
 };
 
 class MaybeEvenBetterCpuPlayer : public BetterCpuPlayer {
